Fix a bunch of warnings due mistmatching types in BOOST_CHECK_EQUAL
This is a typical example:

../../src/test/pow_tests.cpp:30:5: note: in instantiation of function
template specialization
'boost::test_tools::tt_detail::check_frwd<boost::test_tools::tt_detail::equal_impl_frwd,
unsigned int, int>' requested here
    BOOST_CHECK_EQUAL(CalculateNextWorkRequired(&pindexLast,
    nLastRetargetTime, params), 0x1d00d86a);
        ^

+++ src/test/addrman_tests.cpp
@@ -179,7 +179,7 @@ BOOST_AUTO_TEST_CASE(addrman_select)
     {
         ports.insert(addrman.Select().GetPort());
     }
-    BOOST_CHECK_EQUAL(ports.size(), 3);
+    BOOST_CHECK_EQUAL(ports.size(), static_cast<unsigned int>(3));
 }
 
 BOOST_AUTO_TEST_CASE(addrman_new_collisions)
